backend/python_scripts/requirements.txt
# Image processing (use pre-built binary to avoid compilation issues)
pillow>=10.0.0 --only-binary pillow

# Background removal (includes necessary dependencies)
rembg>=2.0.0
